Improve server functions
By default, don't delete keys after batching, just mark as processed.  This helps to ensure we do not accept duplicate keys.

Also added keycount on the diagnosis id so we can track and limit the number of key submissions per uid, to prevent abuse.

+++ ExposureNotification.Backend.Functions/Database/ExposureNotificationStorage.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using ExposureNotification.Backend.Functions;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Newtonsoft.Json;
@@ -14,6 +15,11 @@ namespace ExposureNotification.Backend
 {
 	public class ExposureNotificationStorage
 	{
+		// One person/diagnosis should never need to submit many many keys
+		// and we can detect if they try to and assume it's malicious and prevent it
+		// this is the threshold for how many keys we accept from a single diagnosis uid
+		const int maxKeysPerDiagnosisFile = 30;
+
 		public ExposureNotificationStorage(
 			Action<DbContextOptionsBuilder> buildDbContextOpetions = null,
 			Action<DbContext> initializeDb = null)
@@ -29,7 +35,7 @@ namespace ExposureNotification.Backend
 		readonly DbContextOptions dbContextOptions;
 
 		public Task<List<TemporaryExposureKey>> GetAllKeysAsync()
-        {
+		{
 			using (var ctx = new ExposureNotificationContext(dbContextOptions))
 			{
 				return ctx.TemporaryExposureKeys
@@ -37,7 +43,7 @@ namespace ExposureNotification.Backend
 					.ToListAsync();
 			}
 		}
-			
+
 
 		public void DeleteAllKeysAsync()
 		{
@@ -89,17 +95,32 @@ namespace ExposureNotification.Backend
 		public async Task SubmitPositiveDiagnosisAsync(SelfDiagnosisSubmissionRequest diagnosis)
 		{
 			using (var ctx = new ExposureNotificationContext(dbContextOptions))
+			using (var transaction = ctx.Database.BeginTransaction())
 			{
 				// Ensure the database contains the diagnosis uid
-				if (!ctx.Diagnoses.Any(d => d.DiagnosisUid == diagnosis.DiagnosisUid))
+				var dbDiag = await ctx.Diagnoses.FirstOrDefaultAsync(d => d.DiagnosisUid == diagnosis.DiagnosisUid);
+				
+				// Check that the diagnosis uid exists and that there aren't too many keys associated
+				// already, otherwise it might be someone submitting fake data with a legitimate key
+				if (dbDiag == null || dbDiag.KeyCount > maxKeysPerDiagnosisFile)
 					throw new InvalidOperationException();
 
 				var dbKeys = diagnosis.Keys.Select(k => DbTemporaryExposureKey.FromKey(k)).ToList();
 
+				// Add the new keys to the db
 				foreach (var dbk in dbKeys)
-					ctx.TemporaryExposureKeys.Add(dbk);
+				{
+					// Only add key if it doesn't exist already
+					if (!await ctx.TemporaryExposureKeys.AnyAsync(k => k.Base64KeyData == dbk.Base64KeyData))
+						ctx.TemporaryExposureKeys.Add(dbk);
+				}
+
+				// Increment key count
+				dbDiag.KeyCount += diagnosis.Keys.Count();
 
 				await ctx.SaveChangesAsync();
+
+				await transaction.CommitAsync();
 			}
 		}
 
@@ -113,7 +134,7 @@ namespace ExposureNotification.Backend
 			using (var transaction = ctx.Database.BeginTransaction())
 			{
 				var keys = await ctx.TemporaryExposureKeys
-					.Where(k => k.Region == region)
+					.Where(k => k.Region == region && !k.Processed)
 					.OrderBy(k => k.TimestampMsSinceEpoch)
 					.Take(TemporaryExposureKeyBatches.MaxKeysPerFile)
 					.ToListAsync();
@@ -137,7 +158,17 @@ namespace ExposureNotification.Backend
 
 				await processBatch(f);
 
-				ctx.TemporaryExposureKeys.RemoveRange(keys);
+				// Decide to delete keys or just mark them as processed
+				// Marking as processed is better 
+				if (Startup.DeleteKeysFromDbAfterBatching)
+				{
+					ctx.TemporaryExposureKeys.RemoveRange(keys);
+				}
+				else
+				{
+					foreach (var k in keys)
+						k.Processed = true;
+				}
 
 				await ctx.SaveChangesAsync();
